fix: harden frozen snapshots on Windows
