+++ tests/unit/test_git_subprocess.py
@@ -65,9 +65,14 @@ def test_git_output_rejects_negative_limit() -> None:
 
 def test_git_output_kills_timed_out_process(monkeypatch) -> None:
     proc = _Proc(wait_error=subprocess.TimeoutExpired("git", 1))
-    monkeypatch.setattr(oracle.subprocess, "Popen", lambda *a, **k: proc)
     _error(
-        lambda: oracle.git_output(".", [], deadline=time.monotonic() + 1, limit=1),
+        lambda: bounded.run_git_bounded(
+            ".",
+            [],
+            deadline=time.monotonic() + 1,
+            limit=1,
+            popen=lambda *args, **kwargs: proc,
+        ),
         "DIFF_SNAPSHOT_TIMEOUT",
     )
     assert proc.killed is True
@@ -75,18 +80,28 @@ def test_git_output_kills_timed_out_process(monkeypatch) -> None:
 
 def test_git_output_rejects_nonzero_exit(monkeypatch) -> None:
     proc = _Proc(returncode=3)
-    monkeypatch.setattr(oracle.subprocess, "Popen", lambda *a, **k: proc)
     _error(
-        lambda: oracle.git_output(".", [], deadline=time.monotonic() + 1, limit=1),
+        lambda: bounded.run_git_bounded(
+            ".",
+            [],
+            deadline=time.monotonic() + 1,
+            limit=1,
+            popen=lambda *args, **kwargs: proc,
+        ),
         "DIFF_SNAPSHOT_GIT_ERROR",
     )
 
 
 def test_git_output_rejects_oversize_stdout(monkeypatch) -> None:
     proc = _Proc(out=b"ab")
-    monkeypatch.setattr(oracle.subprocess, "Popen", lambda *a, **k: proc)
     _error(
-        lambda: oracle.git_output(".", [], deadline=time.monotonic() + 1, limit=1),
+        lambda: bounded.run_git_bounded(
+            ".",
+            [],
+            deadline=time.monotonic() + 1,
+            limit=1,
+            popen=lambda *args, **kwargs: proc,
+        ),
         "DIFF_SNAPSHOT_CAPACITY",
     )
     assert proc.killed is True
@@ -99,9 +114,14 @@ def test_git_output_translates_stream_read_error(monkeypatch) -> None:
 
     proc = _Proc()
     proc.stdout = Broken()
-    monkeypatch.setattr(oracle.subprocess, "Popen", lambda *a, **k: proc)
     _error(
-        lambda: oracle.git_output(".", [], deadline=time.monotonic() + 1, limit=1),
+        lambda: bounded.run_git_bounded(
+            ".",
+            [],
+            deadline=time.monotonic() + 1,
+            limit=1,
+            popen=lambda *args, **kwargs: proc,
+        ),
         "DIFF_SNAPSHOT_GIT_ERROR",
     )
 
@@ -109,9 +129,14 @@ def test_git_output_translates_stream_read_error(monkeypatch) -> None:
 def test_git_output_rejects_missing_stdout(monkeypatch) -> None:
     proc = _Proc()
     proc.stdout = None
-    monkeypatch.setattr(oracle.subprocess, "Popen", lambda *a, **k: proc)
     _error(
-        lambda: oracle.git_output(".", [], deadline=time.monotonic() + 1, limit=1),
+        lambda: bounded.run_git_bounded(
+            ".",
+            [],
+            deadline=time.monotonic() + 1,
+            limit=1,
+            popen=lambda *args, **kwargs: proc,
+        ),
         "DIFF_SNAPSHOT_GIT_ERROR",
     )
 
@@ -119,9 +144,14 @@ def test_git_output_rejects_missing_stdout(monkeypatch) -> None:
 def test_git_output_preserves_capacity_error_when_kill_fails(monkeypatch) -> None:
     proc = _Proc(out=b"ab")
     proc.kill = lambda: (_ for _ in ()).throw(OSError())
-    monkeypatch.setattr(oracle.subprocess, "Popen", lambda *a, **k: proc)
     _error(
-        lambda: oracle.git_output(".", [], deadline=time.monotonic() + 1, limit=1),
+        lambda: bounded.run_git_bounded(
+            ".",
+            [],
+            deadline=time.monotonic() + 1,
+            limit=1,
+            popen=lambda *args, **kwargs: proc,
+        ),
         "DIFF_SNAPSHOT_CAPACITY",
     )
 
@@ -142,10 +172,15 @@ def test_git_output_rejects_drain_thread_past_deadline(monkeypatch) -> None:
         def is_alive(self):
             return True
 
-    monkeypatch.setattr(oracle.threading, "Thread", Thread)
-    monkeypatch.setattr(oracle.subprocess, "Popen", lambda *a, **k: proc)
+    monkeypatch.setattr(bounded.threading, "Thread", Thread)
     _error(
-        lambda: oracle.git_output(".", [], deadline=time.monotonic() + 1, limit=1),
+        lambda: bounded.run_git_bounded(
+            ".",
+            [],
+            deadline=time.monotonic() + 1,
+            limit=1,
+            popen=lambda *args, **kwargs: proc,
+        ),
         "DIFF_SNAPSHOT_TIMEOUT",
     )
     assert proc.killed is True
@@ -239,47 +274,6 @@ class _InputProcess:
 
     def kill(self) -> None:
         self.kills += 1
-
-
-def test_windows_process_group_and_taskkill_are_explicit(monkeypatch) -> None:
-    calls = []
-    process = _InputProcess()
-    monkeypatch.setattr(bounded, "_IS_WINDOWS", True)
-    monkeypatch.setattr(bounded, "_TASKKILL", lambda *a, **k: calls.append((a, k)))
-
-    options = bounded._group_options()
-    bounded._kill_group(process)
-
-    assert options == {
-        "creationflags": getattr(subprocess, "CREATE_NEW_PROCESS_GROUP", 0x00000200)
-    }
-    assert calls[0][0][0] == ["taskkill", "/PID", "41", "/T", "/F"]
-    assert process.kills == 0
-
-
-def test_failed_windows_taskkill_falls_back_to_process_kill(monkeypatch) -> None:
-    process = _InputProcess()
-    monkeypatch.setattr(bounded, "_IS_WINDOWS", True)
-    monkeypatch.setattr(
-        bounded,
-        "_TASKKILL",
-        lambda *a, **k: (_ for _ in ()).throw(subprocess.TimeoutExpired("taskkill", 5)),
-    )
-
-    bounded._kill_group(process)
-
-    assert process.kills == 1
-
-
-def test_nonzero_windows_taskkill_falls_back_to_process_kill(monkeypatch) -> None:
-    process = _InputProcess()
-    completed = subprocess.CompletedProcess(["taskkill"], 1)
-    monkeypatch.setattr(bounded, "_IS_WINDOWS", True)
-    monkeypatch.setattr(bounded, "_TASKKILL", lambda *a, **k: completed)
-
-    bounded._kill_group(process)
-
-    assert process.kills == 1
 
 
 def test_nonzero_exit_kills_group_and_reaps_with_exact_waits(monkeypatch) -> None:
@@ -412,49 +406,6 @@ def test_cleanup_thread_join_error_does_not_escape() -> None:
 
 
 @POSIX_SNAPSHOT_TEST
-def test_file_size_limit_uses_single_threaded_exec_guard() -> None:
-    # PR #1252 review thread 4867: threaded parents must not use preexec_fn.
-    proc = _Proc()
-    proc.stdin = io.BytesIO()
-    captured: dict[str, object] = {}
-
-    def popen(command, **kwargs):
-        captured["command"] = command
-        captured.update(kwargs)
-        return proc
-
-    bounded.run_git_bounded(
-        ".",
-        ["hash-object", "-w", "--stdin"],
-        deadline=time.monotonic() + 1,
-        limit=4096,
-        input_=b"payload",
-        popen=popen,
-        file_size_limit=123,
-    )
-
-    command = captured["command"]
-    assert command == [
-        bounded.sys.executable,
-        str(Path(bounded.__file__).with_name("git_exec_guard.py").resolve()),
-        "--fsize",
-        "123",
-        "--",
-        "git",
-        "-c",
-        "core.fsmonitor=false",
-        "-c",
-        command[9],
-        "hash-object",
-        "-w",
-        "--stdin",
-    ]
-    assert str(command[9]).startswith("diff.orderFile=")
-    assert "preexec_fn" not in captured
-    assert captured["start_new_session"] is True
-
-
-@POSIX_SNAPSHOT_TEST
 def test_recursive_object_store_usage_rejects_shared_budget(tmp_path: Path) -> None:
     # PR #1252 review thread 5947: temporary objects consume the shared budget.
     objects = tmp_path / "objects" / "aa"
@@ -467,23 +418,6 @@ def test_recursive_object_store_usage_rejects_shared_budget(tmp_path: Path) -> N
     environment.object_directory = str(tmp_path / "objects")
 
     _error(environment._refresh_object_usage, "DIFF_SNAPSHOT_CAPACITY")
-
-
-def test_file_size_limit_omits_preexec_on_windows(monkeypatch) -> None:
-    # PR #1252 review thread 5947: Windows remains fail-closed upstream.
-    proc = _Proc()
-    captured: dict[str, object] = {}
-    monkeypatch.setattr(bounded, "_IS_WINDOWS", True)
-
-    def popen(*args, **kwargs):
-        captured.update(kwargs)
-        return proc
-
-    bounded.run_git_bounded(
-        ".", [], deadline=time.monotonic() + 1, limit=1, popen=popen, file_size_limit=1
-    )
-
-    assert "preexec_fn" not in captured
 
 
 def test_has_split_index_rejects_missing_entry_terminator() -> None:
@@ -518,10 +452,3 @@ def test_snapshot_git_disables_system_attributes_with_explicit_env() -> None:
         "GIT_ATTR_NOSYSTEM": "1",
         "GIT_NO_REPLACE_OBJECTS": "1",
     }
-
-
-def test_git_exec_guard_rejects_invalid_limit() -> None:
-    # PR #1252 review thread 4867: malformed guards fail before exec.
-    from tree_sitter_analyzer import git_exec_guard
-
-    assert git_exec_guard.main(["--fsize", "bad", "--", "git", "status"]) == 2
