+++ tree_sitter_analyzer/diff_snapshot_paths.py
@@ -2,10 +2,11 @@
 
 from __future__ import annotations
 
-import os
+import json
 from collections.abc import Collection
 
-from .git_path_codec import path_storage
+from .diff_snapshot_capture import FrozenFile
+from .git_path_codec import path_storage, path_to_raw
 from .source_oracle import SourceOracleError, normalize_repo_path
 from .source_oracle_git import GitEpoch
 
@@ -13,6 +14,18 @@ from .source_oracle_git import GitEpoch
 def path_collection_storage(paths: Collection[str]) -> int:
     """Charge retained raw paths and their wire-safe public representations."""
     return sum(path_storage(path) for path in paths)
+
+
+def record_storage(files: tuple[FrozenFile, ...]) -> int:
+    """Charge the deterministic serialized changed-record metadata."""
+    return sum(
+        len(
+            json.dumps(
+                item.record.to_dict(), sort_keys=True, separators=(",", ":")
+            ).encode("utf-8")
+        )
+        for item in files
+    )
 
 
 def epoch_inventory(epoch: GitEpoch, mode: str, limit: int) -> tuple[str, ...]:
@@ -26,7 +39,7 @@ def epoch_inventory(epoch: GitEpoch, mode: str, limit: int) -> tuple[str, ...]:
                 normalize_repo_path(raw.decode("utf-8", "surrogateescape"))
                 for raw in raw_paths
             ),
-            key=os.fsencode,
+            key=path_to_raw,
         )
     )
     if path_collection_storage(paths) > limit:
